REFACTOR: refactoring StackNavigators
