+++ .eslintrc.js
@@ -14,6 +14,7 @@ module.exports = {
     'react/prop-types': 0,
     'global-require': 0,
     '@typescript-eslint/no-use-before-define': 0,
+    'import/prefer-default-export': 0,
   },
   parser: '@typescript-eslint/parser',
   plugins: ['@typescript-eslint', 'import'],
