+++ src/navigation/AndroidTabNavigation/index.tsx
@@ -3,7 +3,8 @@ import { createMaterialBottomTabNavigator } from '@react-navigation/material-bot
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
 import THEME from 'theme';
-import { HomeStackScreen, BookedStackScreen } from 'navigation/StackNavigators';
+import { HomeStackScreen } from 'navigation/StackNavigators/HomeStackScreen';
+import { BookedStackScreen } from 'navigation/StackNavigators/BookedStackScreen';
 
 const AndroidTab = createMaterialBottomTabNavigator();
